fix eslint error
