+++ src/redux/store.js
@@ -1,14 +1,14 @@
-import { configureStore,combineReducers } from "@reduxjs/toolkit";
-import rocketsReducer from "./rockets/rockets";
-import missionsReducer from "./missions/missions";
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
+import rocketsReducer from './rockets/rockets';
+import missionsReducer from './missions/missions';
 
 const rootReducer = combineReducers({
-    rockets: rocketsReducer,
-    missions: missionsReducer,
+  missions: missionsReducer,
+  rockets: rocketsReducer,
 });
 
 const store = configureStore({
-    reducer: rootReducer,
-})
+  reducer: rootReducer,
+});
 
 export default store;
